+++ Changes.md
@@ -59,6 +59,8 @@ v1.5.x is Generally Recommended at this time for new users, has the systems used
 
 Migration Notes:
 
+*   `&mods&` is now a Generalized Placeholder and can only be called within the Presence Formatting GUI
+
 *   `overwriteServerIcon` has been deprecated, instead use the Presence Formatting GUI to define the priority for icons
 
     *   Example: `overwriteServerIcon` = Setting `smallImageKey` to `&PACK&&SERVER&` (Having `&PACK&` be before anything else)